# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API - Autores e Livros",
        "description": "Documentação da API do desafio (Laravel + Sanctum).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/authors": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Listar autores",
                "operationId": "03597029db468f4b49ca68861b2118e2",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Quantidade por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de autores"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Authors"
                ],
                "summary": "Criar autor",
                "operationId": "c58fcfa1a2e042011ee26b57a00e5c2d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Robert C. Martin"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Autor de Clean Code"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Autor criado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/authors/{id}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Mostrar autor",
                "operationId": "a62a46735167db4ec6f25578c0f2ac9c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do autor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Autor encontrado"
                    },
                    "404": {
                        "description": "Autor não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Authors"
                ],
                "summary": "Atualizar autor",
                "operationId": "cf8036fbfd84449d9905eb70a2d63510",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do autor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Robert C. Martin"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Autor de Clean Code (atualizado)"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Autor atualizado com sucesso"
                    },
                    "404": {
                        "description": "Autor não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Authors"
                ],
                "summary": "Deletar autor",
                "operationId": "ea3482a24b060fa981f9cc07c6fa92d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do autor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Autor deletado com sucesso (sem conteúdo)"
                    },
                    "404": {
                        "description": "Autor não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Listar livros",
                "operationId": "921c595372ab399821e267fcf05cbaa1",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Quantidade por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de livros"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Criar livro",
                "operationId": "7ab92284110ace5d0f7de52f372e70c0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "author_id",
                                    "title"
                                ],
                                "properties": {
                                    "author_id": {
                                        "description": "ID do autor (deve existir na tabela authors)",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Clean Code"
                                    },
                                    "year": {
                                        "type": "string",
                                        "example": "2008"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "example": "Programming"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Livro criado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Mostrar detalhes de um livro",
                "operationId": "e079e587c1fee53fe150984526abf3ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do livro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livro encontrado"
                    },
                    "404": {
                        "description": "Livro não encontrado"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Atualizar livro",
                "operationId": "b47b1606570ba86a105bf62f04de6b7f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do livro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "author_id": {
                                        "description": "ID do autor (opcional na atualização, mas se enviado deve existir)",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Clean Code (Atualizado)"
                                    },
                                    "year": {
                                        "type": "string",
                                        "example": "2009"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "example": "Software Engineering"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Livro atualizado com sucesso"
                    },
                    "404": {
                        "description": "Livro não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Deletar livro",
                "operationId": "95c7fc7d78f9dc3884a5bb898b7ead11",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do livro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Livro deletado com sucesso (sem conteúdo)"
                    },
                    "404": {
                        "description": "Livro não encontrado"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authors",
            "description": "Authors"
        },
        {
            "name": "Books",
            "description": "Books"
        }
    ]
}